+++ app/models/AddressesAndUPRN.scala
@@ -18,7 +18,7 @@ package models
 
 import play.api.libs.json.{Json, OFormat}
 
-case class AddressesAndUPRN(address: AddressUk, UPRN: Int)
+case class AddressesAndUPRN(address: AddressUk, UPRN: Long)
 
 object AddressesAndUPRN {
   implicit val format: OFormat[AddressesAndUPRN] = Json.format[AddressesAndUPRN]
